WIP: Discord.ts update, slash commands migration

+++ src/modules/chat_moderation/commands/Clear.ts
@@ -1,19 +1,22 @@
-import {Command, CommandMessage, Guard} from "@typeit/discord";
+import {Discord, Guard, Guild, Slash} from "@typeit/discord";
 import {NotBot} from "../../../guards/NotBot";
 import {HasRoles} from "../../../guards/HasRoles";
 import config from "../../../config.json";
 import {ConfigGuild} from "../../../guards/ConfigGuild";
+import {CommandInteraction} from "discord.js";
 
+@Discord()
 export abstract class Clear {
 
-    @Command("clear")
-    @Guard(NotBot)
-    @Guard(HasRoles([config.ROLES.ADMINISTRATOR]))
-    @Guard(ConfigGuild)
-    private async processCommand(command: CommandMessage) {
+    //@Guard(NotBot)
+    //@Guard(HasRoles([config.ROLES.ADMINISTRATOR]))
+    //@Guard(ConfigGuild)
+    @Guild(config.GUILD)
+    @Slash("clear")
+    private async processCommand(interaction: CommandInteraction) {
 
-        const channel = command.channel;
-        if (!channel.isText()) return;
+        const channel = interaction.channel;
+        if (!channel || !channel.isText()) return;
 
         (await channel.messages.fetch()).forEach((message) => {
             message.delete();
